+++ src/clj/rems/api/services/util.clj
@@ -1,55 +1,54 @@
 (ns rems.api.services.util
   (:require [clojure.test :refer [deftest is testing]]
+            [rems.context :as context]
             [rems.db.roles :as roles]
             [rems.db.users :as users]))
 
-(defn forbidden-organization? [user-id organization]
-  (let [user (users/get-user user-id)
-        not-owner? (not (contains? (roles/get-roles user-id) :owner))
-        incorrect-organization? (not= organization (:organization user))]
-    (when (and not-owner?
-               (or incorrect-organization?
-                   ;; XXX: Special case to forbid an organization owner with
-                   ;;   no organization defined from creating items with
-                   ;;   no organization defined. Consider disallowing undefined
-                   ;;   organizations to be able to remove this.
-                   (and (nil? organization) (nil? (:organization user)))))
-      {:success false
-       :errors [{:type :t.administration.errors/forbidden-organization}]})))
+(defn forbidden-organization? [organization]
+  (let [user-organization (:organization context/*user*)
+        not-owner? (not (contains? context/*roles* :owner))
+        incorrect-organization? (not= organization user-organization)]
+    (and not-owner?
+         (or incorrect-organization?
+             ;; XXX: Special case to forbid an organization owner with
+             ;;   no organization defined from creating or accessing items with
+             ;;   no organization defined. Consider disallowing undefined
+             ;;   organizations to be able to remove this.
+             (and (nil? organization) (nil? user-organization))))))
+
+(defn forbidden-organization-error [organization]
+  (when (forbidden-organization? organization)
+    {:success false
+     :errors [{:type :t.administration.errors/forbidden-organization}]}))
 
 (deftest test-forbidden-organization?
-  (with-redefs [roles/get-roles {"owner" #{:owner}
-                                 "organization-owner" #{:organization-owner}
-                                 "owner-and-organization-owner" #{:owner :organization-owner}
-                                 "organization-owner-with-no-organization" #{:organization-owner}}
-                users/get-user {"owner" {:eppn "owner"
-                                         :organization "own organization"}
-                                "organization-owner" {:eppn "organization-owner"
-                                                      :organization "own organization"}
-                                "owner-and-organization-owner" {:eppn "owner-and-organization-owner"
-                                                                :organization "own organization"}
-                                "organization-owner-with-no-organization" {:eppn "organization-owner-with-no-organization"}}]
-    (let [forbidden-error {:success false
-                           :errors [{:type :t.administration.errors/forbidden-organization}]}]
-      (testing "for owner, all organizations are permitted"
-        (is (nil? (forbidden-organization? "owner" "own organization")))
-        (is (nil? (forbidden-organization? "owner" "not own organization"))))
+  (testing "for owner, all organizations are permitted"
+    (binding [context/*user* {:organization "own organization"}
+              context/*roles* #{:owner}]
+      (is (not (forbidden-organization? "own organization")))
+      (is (not (forbidden-organization? "not own organization")))
+      (is (not (forbidden-organization? "")))
+      (is (not (forbidden-organization? nil)))))
 
-      (testing "for organization owner, only own organization is permitted"
-        (is (nil? (forbidden-organization? "organization-owner" "own organization")))
-        (is (= forbidden-error (forbidden-organization? "organization-owner" "not own organization"))))
+  (testing "for owner who is also an organization owner, all organizations are permitted"
+    (binding [context/*user* {:organization "own organization"}
+              context/*roles* #{:owner :organization-owner}]
+      (is (not (forbidden-organization? "own organization")))
+      (is (not (forbidden-organization? "not own organization")))
+      (is (not (forbidden-organization? "")))
+      (is (not (forbidden-organization? nil)))))
 
-    (testing "for organization owner with no organization, all organizations are forbidden"
-      (is (= forbidden-error (forbidden-organization? "organization-owner-with-no-organization" "own organization"))))
+  (testing "for organization owner, only own organization is permitted"
+    (binding [context/*user* {:organization "own organization"}
+              context/*roles* #{:organization-owner}]
+      (is (not (forbidden-organization? "own organization")))
+      (is (forbidden-organization? "not own organization"))
+      (is (forbidden-organization? ""))
+      (is (forbidden-organization? nil))))
 
-    (testing "for owner who is also an organization owner, all organizations are permitted"
-      (is (nil? (forbidden-organization? "owner-and-organization-owner" "own organization")))
-      (is (nil? (forbidden-organization? "owner-and-organization-owner" "not own organization"))))
-
-    (testing "no organization defined for the created item or organization field left empty"
-      (is (nil? (forbidden-organization? "owner" nil)))
-      (is (nil? (forbidden-organization? "owner" "")))
-      (is (= forbidden-error (forbidden-organization? "organization-owner" nil)))
-      (is (= forbidden-error (forbidden-organization? "organization-owner" "")))
-      (is (= forbidden-error (forbidden-organization? "organization-owner-with-no-organization" nil)))
-      (is (= forbidden-error (forbidden-organization? "organization-owner-with-no-organization" "")))))))
+  (testing "for organization owner with no organization defined, all organizations are forbidden"
+    (binding [context/*user* {}
+              context/*roles* #{:organization-owner}]
+      (is (forbidden-organization? "own organization"))
+      (is (forbidden-organization? ""))
+      (is (forbidden-organization? nil)))))