+++ src/cljs/rems/guide_functions.cljs
@@ -1,5 +1,7 @@
 (ns rems.guide-functions
   (:require [clojure.string :as str]
+            [clojure.test :refer-macros [deftest is]]
+            [rems.common-util :as common-util]
             [rems.git :as git])
   (:require-macros [rems.read-gitlog :refer [read-current-version]]))
 
@@ -8,8 +10,9 @@
                    (apply str (drop 2 line)))))
 
 (defn- link-to-source [meta]
-  (let [link-text (str (:file meta) ":" (:line meta) ":" (:column meta))
-        path (str (:file meta) "#L" (:line meta))
+  (let [file (common-util/normalize-file-path (:file meta))
+        link-text (str file ":" (:line meta) ":" (:column meta))
+        path (str file "#L" (:line meta))
         href (if-let [{:keys [revision]} (read-current-version)]
                (str git/+tree-url+ revision "/" path)
                (str git/+master-url+ path))]