+++ test/clj/rems/api/test_applications.clj
@@ -10,6 +10,7 @@
             [rems.db.test-data :as test-data]
             [rems.handler :refer [handler]]
             [rems.json]
+            [rems.testing-util :refer [with-user]]
             [ring.mock.request :refer :all]))
 
 (use-fixtures
@@ -442,10 +443,16 @@
            (:application/state (get-application application-id user-id))))))
 
 (deftest test-application-submit
-  (let [user-id "alice"
+  (let [owner "owner"
+        user-id "alice"
         form-id (test-data/create-form! {})
         cat-id (test-data/create-catalogue-item! {:form-id form-id})
-        app-id (test-data/create-application! {:catalogue-item-ids [cat-id] :actor user-id})]
+        app-id (test-data/create-application! {:catalogue-item-ids [cat-id] :actor user-id})
+        enable-catalogue-item! #(catalogue/set-catalogue-item-enabled! {:id cat-id
+                                                                        :enabled %})
+        archive-catalogue-item! #(with-user owner
+                                   (catalogue/set-catalogue-item-archived! {:id cat-id
+                                                                            :archived %}))]
     (testing "submit with disabled catalogue item fails"
       (is (:success (catalogue/set-catalogue-item-enabled! {:id cat-id
                                                             :enabled false})))
@@ -455,20 +462,16 @@
              (send-command user-id {:type :application.command/submit
                                     :application-id app-id}))))
     (testing "submit with archived catalogue item fails"
-      (is (:success (catalogue/set-catalogue-item-enabled! {:id cat-id
-                                                            :enabled true})))
-      (is (:success (catalogue/set-catalogue-item-archived! {:id cat-id
-                                                             :archived true})))
+      (is (:success (enable-catalogue-item! true)))
+      (is (:success (archive-catalogue-item! true)))
       (rems.db.applications/reload-cache!)
       (is (= {:success false
               :errors [{:type "t.actions.errors/disabled-catalogue-item" :catalogue-item-id cat-id}]}
              (send-command user-id {:type :application.command/submit
                                     :application-id app-id}))))
     (testing "submit with normal catalogue item succeeds"
-      (is (:success (catalogue/set-catalogue-item-enabled! {:id cat-id
-                                                            :enabled true})))
-      (is (:success (catalogue/set-catalogue-item-archived! {:id cat-id
-                                                             :archived false})))
+      (is (:success (enable-catalogue-item! true)))
+      (is (:success (archive-catalogue-item! false)))
       (rems.db.applications/reload-cache!)
       (is (= {:success true}
              (send-command user-id {:type :application.command/submit
@@ -519,16 +522,54 @@
         cat-id (test-data/create-catalogue-item! {:workflow-id wf-id})
         app-id (test-data/create-application! {:catalogue-item-ids [cat-id]
                                                :actor applicant})]
+    (testing "applicant's commands for draft"
+      (is (= #{"application.command/accept-licenses"
+               "application.command/change-resources"
+               "application.command/close"
+               "application.command/copy-as-new"
+               "application.command/invite-member"
+               "application.command/remove-member"
+               "application.command/save-draft"
+               "application.command/submit"
+               "application.command/uninvite-member"}
+             (set (:application/permissions (get-application app-id applicant))))))
     (testing "submit"
       (is (= {:success true}
              (send-command applicant {:type :application.command/submit
                                       :application-id app-id}))))
+    (testing "applicant's commands after submit"
+      (is (= #{"application.command/accept-licenses"
+               "application.command/copy-as-new"
+               "application.command/remove-member"
+               "application.command/uninvite-member"}
+             (set (:application/permissions (get-application app-id applicant))))))
+    (testing "handler's commands"
+      (is (= #{"application.command/add-licenses"
+               "application.command/add-member"
+               "application.command/assign-external-id"
+               "application.command/change-resources"
+               "application.command/close"
+               "application.command/invite-member"
+               "application.command/remark"
+               "application.command/remove-member"
+               "application.command/request-decision"
+               "application.command/request-review"
+               "application.command/return"
+               "application.command/uninvite-member"
+               "see-everything"}
+             (set (:application/permissions (get-application app-id handler))))))
     (testing "request decision"
       (is (= {:success true}
              (send-command handler {:type :application.command/request-decision
                                     :application-id app-id
                                     :deciders [decider]
                                     :comment ""}))))
+    (testing "decider's commands"
+      (is (= #{"application.command/approve"
+               "application.command/reject"
+               "application.command/remark"
+               "see-everything"}
+             (set (:application/permissions (get-application app-id decider))))))
     (testing "approve"
       (is (= {:success true}
              (send-command decider {:type :application.command/approve